Refactor away dead "forwarded whitelist" code
